+++ src/main/java/alexthw/not_enough_glyphs/datagen/DatagenMain.java
@@ -30,7 +30,7 @@ public class DatagenMain {
         generator.addProvider(event.includeClient(), new NEGItemModels(generator, fileHelper));
         generator.addProvider(event.includeServer(), new NEGGlyphRecipeProvider(generator));
         generator.addProvider(event.includeServer(), new NEGApparatusProvider(generator));
-
+        generator.addProvider(event.includeServer(), new NEGEntityTagProvider(generator, provider, fileHelper));
         generator.addProvider(event.includeClient(), new NEGLangProvider(generator.getPackOutput(), NotEnoughGlyphs.MODID, "en_us"));
     }